Implement ActivitySubtype base class in existing tests
