+++ api-rework/invasives/api/tests/subtypes/test_aquatic_mechanical_treatment.py
@@ -1,10 +1,7 @@
-from django.test import TestCase
-from django.test.client import Client
-from api.models.activity.activity import Activity
-from api.serializers.activity import ActivitySerializer
+from .base import BaseActivitySubtypeTest
 
 
-class AquaticMechanicalTreatmentTest(TestCase):
+class AquaticMechanicalTreatmentTest(BaseActivitySubtypeTest):
 
     fixtures = [
         "test/subtypes/treatments/test_aquatic_mechanical_treatment_codes",
@@ -12,52 +9,18 @@ class AquaticMechanicalTreatmentTest(TestCase):
     ]
 
     def test_expect_two_activities(self):
-        """Verify that two activities are returned from the list endpoint"""
-        client = Client()
-        result = client.get("/activities")
-        self.assertEqual(result.status_code, 200)
-
-        response_object = result.json()
-        self.assertEqual(len(response_object), 2)
+        self.expect_two_activities()
 
     def test_no_pac_number_present(self):
-        """Check that participants do not include a pac number when fetched for an Observation activity"""
-        try:
-            record = Activity.objects.get(id="6BBA2749-EE3D-41B6-A9F1-4A0CB37029F7")
-            self.assertIsNotNone(record)
-
-            serial = ActivitySerializer(record)
-
-            self.assertIsNotNone(serial.data["subtype_data"]["mechanical_treatments"])
-            serial.data["participants"]
-
-            for person in serial.data["participants"]:
-                self.assertIsNotNone(person.get("name"))
-                self.assertIsNone(person.get("pac_number"))
-
-        except Activity.DoesNotExist:
-            self.fail("Activity does not exist")
+        self.no_pac_number_present()
 
     def test_casting_fixture_into_serializer(self):
-        """fetching an activity can be cast into serializer"""
-        try:
-            record = Activity.objects.get(id="6BBA2749-EE3D-41B6-A9F1-4A0CB37029F7")
-            self.assertIsNotNone(record)
-
-            serial = ActivitySerializer(record)
-            # Check a subtype specific field
-            self.assertIsNotNone(serial.data["subtype_data"]["mechanical_treatments"])
-
-        except Activity.DoesNotExist:
-            self.fail("Activity does not exist")
+        self.casting_fixture_into_serializer(expected_subtype_key="mechanical_treatments")
 
     def test_subtype_details_full(self):
         """Subtype keys match the information from fixtures."""
-        client = Client()
-        result = client.get("/activities/6BBA2749-EE3D-41B6-A9F1-4A0CB37029F7")
-        self.assertEqual(result.status_code, 200)
+        response_object = self.fetch_a().json()
 
-        response_object = result.json()
         self.assertEqual(
             len(response_object["subtype_data"]["mechanical_treatments"]), 1
         )
@@ -82,11 +45,7 @@ class AquaticMechanicalTreatmentTest(TestCase):
 
     def test_subtype_details_two_entries(self):
         """Subtype keys match the information from fixtures."""
-        client = Client()
-        result = client.get("/activities/CD542709-F767-402F-818E-117B3FBC797D")
-        self.assertEqual(result.status_code, 200)
-
-        response_object = result.json()
+        response_object = self.fetch_b().json()
         sd = response_object["subtype_data"]
         self.assertEqual(sd["authorization_info"], "The test looks for this")
