+++ api-rework/invasives/api/tests/subtypes/test_biocontrol_collection.py
@@ -1,70 +1,21 @@
-from django.test import TestCase
-from django.test.client import Client
-from api.models.activity import Activity
-from api.serializers.activity import ActivitySerializer
+from .base import BaseActivitySubtypeTest
 
 
-class BiocontrolReleaseTest(TestCase):
+class BiocontrolReleaseTest(BaseActivitySubtypeTest):
 
   fixtures = [
     "test/subtypes/biocontrol/test_biocontrol_collection_codes",
     "test/subtypes/biocontrol/test_biocontrol_collection",
   ]
 
-  def fetch(self, id):
-    client = Client()
-
-    result = client.get(f"/activities/{id}")
-    self.assertEqual(result.status_code, 200)
-    return result
-
-  def fetch_a(self):
-    return self.fetch("6BBA2749-EE3D-41B6-A9F1-4A0CB37029F7")
-
-  def fetch_b(self):
-    return self.fetch("CD542709-F767-402F-818E-117B3FBC797D")
-
   def test_expect_two_activities(self):
-    """Verify that two activities are returned from the list endpoint"""
-    client = Client()
-    result = client.get("/activities")
-    self.assertEqual(result.status_code, 200)
-
-    response_object = result.json()
-    self.assertEqual(len(response_object), 2)
+      self.expect_two_activities()
 
   def test_no_pac_number_present(self):
-    """Check that participants do not include a pac number when fetched for an Observation activity"""
-    try:
-        record = Activity.objects.get(id="6BBA2749-EE3D-41B6-A9F1-4A0CB37029F7")
-        self.assertIsNotNone(record)
-
-        serial = ActivitySerializer(record)
-
-        self.assertIsNotNone(
-            serial.data["subtype_data"]["collection_information"]
-        )
-
-        for person in serial.data["participants"]:
-            self.assertIsNotNone(person.get("name"))
-            self.assertIsNone(person.get("pac_number"))
-
-    except Activity.DoesNotExist:
-        self.fail("Activity does not exist")
+      self.no_pac_number_present()
 
   def test_casting_fixture_into_serializer(self):
-    """fetching an activity can be cast into serializer"""
-    try:
-      record = Activity.objects.get(id="CD542709-F767-402F-818E-117B3FBC797D")
-      self.assertIsNotNone(record)
-
-      serial = ActivitySerializer(record)
-      # Check a subtype specific field
-      self.assertGreaterEqual(
-          len(serial.data["subtype_data"]["collection_information"]), 1
-      )
-    except Activity.DoesNotExist:
-      self.fail("Activity does not exist")
+      self.casting_fixture_into_serializer(expected_subtype_key="collection_information")
 
   def test_subtype_details_full(self):
     """Subtype keys match the information from fixtures."""
